+++ gnp/src/main/java/com/gnp/auth/controllers/TestController.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package com.gnp.auth.controllers;
 
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
