+++ gnp/src/main/java/com/gnp/auth/payload/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package com.gnp.auth.payload.request;
 
 import java.util.Set;
