+++ gnp/src/main/java/com/gnp/auth/repository/RoleRepository.java
@@ -1,9 +1,12 @@
-package com.gnp.auth;
+package com.gnp.auth.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.gnp.auth.models.ERole;
+import com.gnp.auth.models.Role;
 
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
