+++ gnp/src/main/java/com/gnp/auth/repository/UserRepository.java
@@ -1,7 +1,9 @@
-package com.gnp.auth;
+package com.gnp.auth.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.gnp.auth.models.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
