+++ gnp/src/main/java/com/gnp/auth/security/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package com.gnp.auth.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -13,6 +13,10 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import com.gnp.auth.security.jwt.AuthEntryPointJwt;
+import com.gnp.auth.security.jwt.AuthTokenFilter;
+import com.gnp.auth.security.services.UserDetailsServiceImpl;
 
 @Configuration
 @EnableWebSecurity
