+++ gnp/src/main/java/com/gnp/auth/security/jwt/AuthTokenFilter.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package com.gnp.auth.security.jwt;
 
 import java.io.IOException;
 
@@ -16,6 +16,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
+
+import com.gnp.auth.security.services.UserDetailsServiceImpl;
 
 public class AuthTokenFilter extends OncePerRequestFilter {
 	@Autowired
