+++ gnp/src/main/java/com/gnp/auth/security/services/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package com.gnp.auth.security.services;
 
 import java.util.Collection;
 import java.util.List;
@@ -10,6 +10,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.gnp.auth.models.User;
 
 public class UserDetailsImpl implements UserDetails {
 	private static final long serialVersionUID = 1L;
