+++ gnp/src/main/java/com/gnp/auth/security/services/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.gnp.auth;
+package com.gnp.auth.security.services;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +7,9 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import com.gnp.auth.models.User;
+import com.gnp.auth.repository.UserRepository;
 
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
